Put a couple of source files into a sub-package before splitting them up.
