+++ core/src/main/scala/com/rackspace/com/papi/components/checker/handler/NullHandler.scala
@@ -15,7 +15,8 @@
  */
 package com.rackspace.com.papi.components.checker.handler
 
-import com.rackspace.com.papi.components.checker.step.{Step, StepContext, Result}
+import com.rackspace.com.papi.components.checker.step.results.Result
+import com.rackspace.com.papi.components.checker.step.base.Step
 import com.rackspace.com.papi.components.checker.servlet._
 
 import com.rackspace.com.papi.components.checker.Validator
