+++ core/src/main/scala/com/rackspace/com/papi/components/checker/handler/SaveDotHandler.scala
@@ -17,10 +17,10 @@ package com.rackspace.com.papi.components.checker.handler
 
 import java.io.File
 
+import com.rackspace.com.papi.components.checker.step.results.Result
 import com.rackspace.com.papi.components.checker.wadl.WADLDotBuilder
 import com.rackspace.com.papi.components.checker._
 import com.rackspace.com.papi.components.checker.servlet._
-import com.rackspace.com.papi.components.checker.step.Result
 
 import javax.xml.transform.dom.DOMSource
 import javax.xml.transform.stream.StreamResult
