+++ core/src/main/scala/com/rackspace/com/papi/components/checker/handler/ServletResultHandler.scala
@@ -17,9 +17,7 @@ package com.rackspace.com.papi.components.checker.handler
 
 import com.rackspace.com.papi.components.checker._
 import com.rackspace.com.papi.components.checker.servlet._
-import com.rackspace.com.papi.components.checker.step.Result
-import com.rackspace.com.papi.components.checker.step.ErrorResult
-import com.rackspace.com.papi.components.checker.step.MultiFailResult
+import com.rackspace.com.papi.components.checker.step.results.{MultiFailResult, ErrorResult, Result}
 
 import scala.collection.JavaConversions._
