+++ core/src/main/scala/com/rackspace/com/papi/components/checker/step/URIXSD.scala
@@ -21,6 +21,9 @@ import javax.xml.validation.Schema
 import com.rackspace.com.papi.components.checker.servlet._
 import javax.servlet.FilterChain
 
+import com.rackspace.com.papi.components.checker.step.base.{ConnectedStep, Step, StepContext}
+import com.rackspace.com.papi.components.checker.step.results.{MultiFailResult, MismatchResult, Result}
+
 class URIXSD(id : String, label : String, val simpleType : QName, val schema : Schema, val captureHeader : Option[String], next : Array[Step])
       extends ConnectedStep(id, label, next) {
