+++ core/src/main/scala/com/rackspace/com/papi/components/checker/step/startend/start-end.scala
@@ -13,19 +13,19 @@
  *   See the License for the specific language governing permissions and
  *   limitations under the License.
  */
-package com.rackspace.com.papi.components.checker.step
+package com.rackspace.com.papi.components.checker.step.startend
 
 import java.util.HashMap
-
-import scala.util.control.Breaks._
-
+import javax.servlet.FilterChain
 import javax.xml.namespace.QName
 import javax.xml.validation.Schema
 
-import scala.util.matching.Regex
-
 import com.rackspace.com.papi.components.checker.servlet._
-import javax.servlet.FilterChain
+import com.rackspace.com.papi.components.checker.step.XSDStringValidator
+import com.rackspace.com.papi.components.checker.step.base.{ConnectedStep, Step, StepContext}
+import com.rackspace.com.papi.components.checker.step.results._
+
+import scala.util.matching.Regex
 
 //
 //  The start step
