+++ core/src/test/scala/com/rackspace/com/papi/components/checker/handler/MethodLabelHandlerTest.scala
@@ -25,6 +25,9 @@ import javax.xml.transform.stream.StreamSource
 import com.rackspace.com.papi.components.checker.Converters._
 import com.rackspace.com.papi.components.checker.servlet.{CheckerServletRequest, CheckerServletResponse}
 import com.rackspace.com.papi.components.checker.step._
+import com.rackspace.com.papi.components.checker.step.base.StepContext
+import com.rackspace.com.papi.components.checker.step.results.Result
+import com.rackspace.com.papi.components.checker.step.startend.Accept
 import com.rackspace.com.papi.components.checker.util.{IdentityTransformPool, HeaderMap}
 import com.rackspace.com.papi.components.checker.{AssertResultHandler, TestConfig, BaseValidatorSuite, Validator}
 import org.junit.runner.RunWith
