+++ core/src/test/scala/com/rackspace/com/papi/components/checker/wadl/BaseStepSpec.scala
@@ -15,6 +15,9 @@
  */
 package com.rackspace.com.papi.components.checker.wadl
 
+import com.rackspace.com.papi.components.checker.step.base.{ConnectedStep, Step}
+import com.rackspace.com.papi.components.checker.step.startend._
+
 import scala.collection.mutable.LinkedList
 
 import org.scalatest.exceptions.TestFailedException
